backend/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
transformers==4.36.0
torch==2.1.0
librosa==0.10.1
numpy==1.24.3
requests==2.31.0
gunicorn==21.2.0
whitenoise==6.5.0
python-dotenv==1.0.0
dj-database-url==2.1.0
psycopg2-binary==2.9.7
PyPDF2==3.0.1
python-docx==1.1.0